WareHouse/ProjectMASUI-2_DefaultOrganization_20240111221522/cloud_storage.py
from google.cloud import storage
from onedrive import OneDrive
# Add the necessary code for Google Cloud and OneDrive integration
# Implement the necessary code for Google Cloud and OneDrive integration
class GoogleCloudStorage:
    def __init__(self):
        self.storage = storage.Client()
    def authenticate(self, credentials):
        # Authenticate with Google Cloud using the provided credentials
        pass
    def upload_file(self, file_path, bucket_name):
        # Upload a file to the specified Google Cloud Storage bucket
        pass
    def download_file(self, file_path, bucket_name):
        # Download a file from the specified Google Cloud Storage bucket
        pass
class OneDriveStorage:
    def __init__(self):
        self.onedrive = OneDrive()
    def authenticate(self, credentials):
        # Authenticate with OneDrive using the provided credentials
        pass
    def upload_file(self, file_path, folder_id):
        # Upload a file to the specified OneDrive folder
        pass
    def download_file(self, file_path, folder_id):
        # Download a file from the specified OneDrive folder
        pass