WareHouse/ProjectMASUI-2_DefaultOrganization_20240111221522/frontend.py
'''
This module contains the frontend UI for the ProjectMASUI application.
'''
import React from 'react';
class MASInteractionsSection extends React.Component {
    render() {
        return (
            <div>
                {/* Placeholder for MAS interactions */}
            </div>
        );
    }
}
class LLMManagementSection extends React.Component {
    render() {
        return (
            <div>
                {/* Placeholder for LLM management */}
            </div>
        );
    }
}
class CloudStorageIntegrationSection extends React.Component {
    render() {
        return (
            <div>
                {/* Placeholder for cloud storage integration */}
            </div>
        );
    }
}
class App extends React.Component {
    render() {
        return (
            <div>
                <MASInteractionsSection />
                <LLMManagementSection />
                <CloudStorageIntegrationSection />
            </div>
        );
    }
}
export default App;