WareHouse/ProjectMASUI-2_DefaultOrganization_20240111221522/requirements.txt
Django==3.2.7
Flask==2.0.1
google-cloud-storage==1.42.0
pandas==1.3.3
numpy==1.21.2
